% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd_data_functions.R
\name{download_nlcd_chunks}
\alias{download_nlcd_chunks}
\title{download all chunks needed for nlcd multiple cell numbers ahead of time}
\usage{
download_nlcd_chunks(nlcd_cell_numbers)
}
\arguments{
\item{nlcd_cell_numbers}{vector of nlcd cell numbers}
}
\value{
downloaded fst files in nlcd_fst folder in working directory
}
\description{
download all chunks needed for nlcd multiple cell numbers ahead of time
}
\examples{
if (FALSE) {
nlcd_cell_numbers <- c(7814606790, 7814606790, 7756256174)

download_nlcd_chunks(nlcd_cell_numbers)
}
}
