% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd_data_functions.R
\name{get_nlcd_cell_numbers_points}
\alias{get_nlcd_cell_numbers_points}
\title{get NLCD cell numbers for given lat and lon}
\usage{
get_nlcd_cell_numbers_points(point_data)
}
\arguments{
\item{point_data}{data.frame with columns 'lat' and 'lon'}
}
\value{
a data.frame identical to the input data.frame but with appended NLCD cell numbers
}
\description{
get NLCD cell numbers for given lat and lon
}
\examples{
if (FALSE) {
point_data <- data.frame(
  id = c('1a', '2b', '3c'),
  lat = c(39.19674, 39.19674, 39.28765),
  lon = c(-84.582601, -84.582601, -84.510173)
)

get_nlcd_cell_numbers_points(point_data)
}
}
