% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd_polygon_functions.R
\name{get_nlcd_data_point_buffer}
\alias{get_nlcd_data_point_buffer}
\title{get NLCD data for specified buffer radius around point data}
\usage{
get_nlcd_data_point_buffer(point_data, buffer_m)
}
\arguments{
\item{point_data}{data.frame with columns 'lat' and 'lon'}

\item{buffer_m}{desired buffer radius in meters}
}
\value{
a data.frame identical to the input data.frame but with appended average NLCD values (and in long format)
all available products and years will be returned.
}
\description{
get NLCD data for specified buffer radius around point data
}
\examples{
if (FALSE) {
point_data <- data.frame(
  id = c('1a', '2b', '3c'),
  lat = c(39.19674, 39.19674, 39.28765),
  lon = c(-84.582601, -84.582601, -84.510173)
)

get_nlcd_data_polygons(point_data, buffer_m = 400)
}
}
