% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcd_polygon_functions.R
\name{get_nlcd_data_polygons}
\alias{get_nlcd_data_polygons}
\title{get NLCD data for polygons}
\usage{
get_nlcd_data_polygons(polygon_data)
}
\arguments{
\item{polygon_data}{an sf data.frame containing polygons for which data from nlcd cells will be averaged}
}
\value{
a data.frame identical to the input data.frame but with appended percentage NLCD values (and in long format)
all available products and years will be returned.
}
\description{
get NLCD data for polygons
}
